requirements.txt
websockets>=12.0
sounddevice>=0.4
numpy>=1.24
google-genai>=1.0
httpx>=0.27
python-dotenv>=1.0
rich>=13.0
